Update components.md -- "Classic MVC" example
@ArthurClements pointed out in [this gitter comment](https://gitter.im/lhorie/mithril.js?at=55664ba1d21e5ed02ff06e54) that the Classic MVC example is broken.

I wanted to show a working jsfiddle with these changes applied, but I couldn't figure out how to get a POST request working. I do, however, have a jsbin that is working, although there are significant changes in the code to allow for GET and POST.
http://jsbin.com/cokesu/2/edit?js,output

I don't know why there is a redraw called for every character entered. It's as though `oninput` is really `onchange`