add chat and wiki to README
I don't know about you, but when I hit a repo, I read the README and dive into the code. I never think to check the wiki. Newbies in gitter chat never seem to know where to look for more info than the website provides. I wonder if the wiki should be transferred to a dedicated page on the website?